# apiSwagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Movie API",
      "description": "API for managing movies and users in a movie database.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080"
      }
    ],
    "tags": [
        {
            "name": "Main"
        },
        {
            "name": "Movies",
            "description": "Operations related to movies collection"
        },
        {
            "name": "Users",
            "description": "Operations related to users collection"
        }
    ],
    "paths": {
      "/": {
        "get": {
          "tags": ["Main"], 
          "summary": "Welcome Message",
          "responses": {
            "200": {
              "description": "Welcome message",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/movies": {
        "get": {
          "tags": ["Movies"], 
          "summary": "Retrieve all movies",
          "responses": {
            "200": {
              "description": "A list of movies",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "genre": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "director": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/movies/{title}": {
        "get": {
          "tags": ["Movies"],
          "summary": "Retrieve a specific movie by title",
          "parameters": [
            {
              "name": "title",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "genre": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "director": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Movie not found"
            }
          }
        }
      },
      "/movies/genres/{genreName}": {
        "get": {
          "tags": ["Movies"],
          "summary": "Retrieve genre by name",
          "parameters": [
            {
              "name": "genreName",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Genre found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "genre": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Genre not found"
            }
          }
        }
      },
      "/movies/directors/{directorName}": {
        "get": {
          "tags": ["Movies"],
          "summary": "Retrieve director by name",
          "parameters": [
            {
              "name": "directorName",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Director found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "director": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Director not found"
            }
          }
        }
      },
      "/users": {
        "post": {
          "tags": ["Users"],
          "summary": "Create a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Username": {
                      "type": "string"
                    },
                    "Password": {
                      "type": "string"
                    },
                    "Email": {
                      "type": "string"
                    },
                    "Birthday": {
                      "type": "string",
                      "format": "date"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created"
            },
            "400": {
              "description": "User already exists"
            }
          }
        }
      },
      "/users/{Username}": {
        "put": {
          "tags": ["Users"],
          "summary": "Update user by username",
          "parameters": [
            {
              "name": "Username",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Username": {
                      "type": "string"
                    },
                    "Password": {
                      "type": "string"
                    },
                    "Email": {
                      "type": "string"
                    },
                    "Birthday": {
                      "type": "string",
                      "format": "date"
                    }
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "User updated"
              },
              "404": {
                "description": "User not found"
              }
            }
          }
        },
        "delete": {
          "tags": ["Users"],
          "summary": "Delete user by username",
          "parameters": [
            {
              "name": "Username",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User deleted"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
      "/users/{Username}/movies/{movieID}": {
        "post": {
          "tags": ["Users"],
          "summary": "Add a movie to user's favorites",
          "parameters": [
            {
              "name": "Username",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "movieID",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie added to favorites"
            },
            "404": {
              "description": "User or movie not found"
            }
          }
        },
        "delete": {
          "tags": ["Users"],
          "summary": "Remove a movie from user's favorites",
          "parameters": [
            {
              "name": "Username",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "movieID",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movie removed from favorites"
            },
            "404": {
              "description": "User or movie not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "Username": {
              "type": "string"
            },
            "Password": {
              "type": "string"
            },
            "Email": {
              "type": "string"
            },
            "Birthday": {
              "type": "string",
              "format": "date"
            }
          }
        },
        "Movie": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "genre": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "director": {
              "type": "string"
            }
          }
        }
      }
    }
  }
  